build: fix viz tests
removed test_precomp.cpp
